requirements.txt
feedparser>=6.0.11
feedgen>=1.0.0
requests>=2.32.3
beautifulsoup4>=4.12.3
pydantic>=2.11.0
PyYAML>=6.0.2
python-dateutil>=2.9.0.post0
pytest>=8.3.2
